refactor: 당첨 금액 합산 결과가 Integer의 범위를 벗어날 경우를 대비하여 자료형을 long으로 수정

+++ src/main/java/lotto/model/WinningResult.java
@@ -20,9 +20,9 @@ public class WinningResult {
         return rankCount.getOrDefault(lottoRank, 0);
     }
 
-    public Integer calculateTotalPrize() {
+    public long calculateTotalPrize() {
         return rankCount.entrySet().stream()
-                .mapToInt(entry -> entry.getKey().getPrize() * entry.getValue())
+                .mapToLong(entry -> (long) entry.getKey().getPrize() * entry.getValue())
                 .sum();
     }
 }
